artifacts/api_server/requirements.txt
discord.py>=2.3.0
asyncpg>=0.29.0
google-generativeai>=0.8.0
aiohttp>=3.9.0
firebase-admin>=6.0.0
